feat(sync): SyncProtocolExecutor trait for unified protocol testing (#1972)

+++ crates/node/primitives/src/sync/wire.rs
@@ -71,7 +71,13 @@ pub enum StreamMessage<'a> {
     /// Follow-up message in an ongoing sync operation.
     Message {
         /// Sequence number for ordering.
-        sequence_id: usize,
+        ///
+        /// # Wire Format Change
+        ///
+        /// Changed from `usize` to `u64` for cross-platform portability.
+        /// This is a breaking wire format change - nodes must be upgraded
+        /// together to avoid deserialization failures.
+        sequence_id: u64,
         /// The message payload.
         payload: MessagePayload<'a>,
         /// Nonce for the next message.
